frames/requirements.txt
aiofiles==22.1.0
docker==6.0.0
asyncpg
databases==0.6.1
fastapi==0.85.0
Jinja2==3.1.2
ormar==0.11.3
requests==2.28.1
pytest==6.2.4
pydantic[email]==1.10.2
python-dotenv==0.21.0
python-multipart==0.0.5
psycopg2
sqlalchemy==1.4.41
uvicorn==0.18.3
bcrypt==4.0.0
passlib==1.7.4
python-jose==3.3.0
aiosqlite==0.17.0
alembic==1.8.1